local-first-analytics-stack/pyproject.toml
[project]
name = "local-first-analytics-stack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dlt[ducklake]>=1.20.0",
    "python-dotenv>=1.2.1",
    "sqlmesh[duckdb]>=0.227.1",
]
